requirements.txt
asgiref==3.8.1
black==24.4.1
certifi==2025.1.31
charset-normalizer==3.4.1
click==8.1.8
dj-database-url==2.3.0
Django==4.2.17
django-filter==24.3
django-modeltranslation==0.19.12
djangorestframework==3.15.2
djangorestframework_simplejwt==5.3.1
drf-yasg==1.21.10
idna==3.10
inflection==0.5.1
mypy-extensions==1.0.0
packaging==24.2
pathspec==0.12.1
platformdirs==4.3.6
psycopg==3.2.6
psycopg2-binary==2.9.10
PyJWT==2.9.0
python-dotenv==1.0.1
pytz==2025.1
PyYAML==6.0.2
requests==2.32.3
sqlparse==0.5.3
typing_extensions==4.12.2
uritemplate==4.1.1
urllib3==2.2.3
